Fixed errors caused by float16/float32 type issues in whisper/test.py and whisper/benchmark.py.

+++ whisper/test.py
@@ -5,13 +5,11 @@ import unittest
 import mlx.core as mx
 import numpy as np
 import os
-import subprocess
 import torch
 
 import whisper
 import whisper.audio as audio
 import whisper.load_models as load_models
-import whisper.torch_whisper as torch_whisper
 import whisper.decoding as decoding
 
 
@@ -62,7 +60,7 @@ class TestWhisper(unittest.TestCase):
         dims = mlx_model.dims
         mels = mx.array(np.random.randn(1, 3_000, dims.n_mels), mx.float16)
         tokens = mx.array(np.random.randint(0, dims.n_vocab, (1, 20)), mx.int32)
-        logits = mlx_model(mels, tokens)
+        logits = mlx_model(mels, tokens).astype(mx.float16)
         self.assertEqual(logits.dtype, mx.float16)
